EdgeSolution/modules/pyFileManagerModule/requirements.txt
azure-iot-device~=2.5.1
azure-storage-blob